# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smakvärlden API — recipe costs, ingredients, and chef statistics
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered and token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or email taken
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
...
  /ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: List all ingredients
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Create an ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngredientBody"
      responses:
        "201":
          description: Created ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
  /ingredients/{id}:
    get:
      operationId: getIngredient
      tags: [ingredients]
      summary: Get ingredient by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
    put:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Update an ingredient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngredientBody"
      responses:
        "200":
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete an ingredient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /community/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      summary: List community posts / shared recipes
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Share a recipe to the community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommunityPostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}/like:
    post:
      operationId: likeCommunityPost
      tags: [community]
      summary: Like a community post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated like count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with the AI chef assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/suggest:
    post:
      operationId: aiSuggestRecipe
      tags: [ai]
      summary: AI recipe suggestions based on ingredients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSuggestBody"
      responses:
        "200":
          description: AI suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    Ingredient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentPriceSek:
          type: number
        priceChangePct:
          type: number
        supplier:
          type: string
        updatedAt:
          type: string
          format: date-time
      required: [id, name, category, unit, currentPriceSek, priceChangePct, updatedAt]
    CreateIngredientBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentPriceSek:
          type: number
        supplier:
          type: string
      required: [name, category, unit, currentPriceSek]
...
    CommunityPost:
      type: object
      properties:
        id:
          type: integer
        recipeName:
          type: string
        chefName:
          type: string
        description:
          type: string
        category:
          type: string
        costSek:
          type: number
        likes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, recipeName, chefName, description, category, costSek, likes, createdAt]
    CreateCommunityPostBody:
      type: object
      properties:
        recipeName:
          type: string
        chefName:
          type: string
        description:
          type: string
        category:
          type: string
        costSek:
          type: number
      required: [recipeName, chefName, description, category, costSek]
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
            required: [role, content]
      required: [message]
    AiSuggestBody:
      type: object
      properties:
        ingredients:
          type: array
          items:
            type: string
        budget:
          type: number
        servings:
          type: integer
      required: [ingredients]
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        tokensUsed:
          type: integer
      required: [reply]